requirements.txt
addict==2.4.0
cfgv==3.3.1
distlib==0.3.6
filelock==3.9.0
identify==2.5.13
nodeenv==1.7.0
platformdirs==2.6.2
pre-commit==2.21.0
python-dotenv==0.21.0
PyYAML==6.0
virtualenv==20.17.1
pygame==2.1.2
pymunk==6.4.0
numpy==1.24.2
pathfinding==1.0.1
